+++ lib/provider/map_notifier.dart
@@ -1,6 +1,13 @@
+import 'dart:async';
+
+import 'package:concordi_around/models/coordinate.dart';
 import 'package:flutter/widgets.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 class MapNotifier with ChangeNotifier {
+
+  Completer<GoogleMapController> _completer = Completer();
+  Completer<GoogleMapController> get getCompleter {return _completer;}
   bool showFloorPlan = false;
   bool showEnterBuilding = false;
   int selectedFloorPlan = 8;
@@ -18,5 +25,24 @@ class MapNotifier with ChangeNotifier {
   void setEnterBuildingVisibility(bool visibility) {
     showEnterBuilding = visibility;
     notifyListeners();
+  }
+
+  Future<void> goToSpecifiedLatLng(Coordinate coordinate) async {
+    final GoogleMapController controller = await _completer.future;
+    if (coordinate != null) {
+      CameraPosition _newPosition = CameraPosition(target: coordinate.toLatLng(), zoom: 18.5);
+      controller.animateCamera(CameraUpdate.newCameraPosition(_newPosition));
+    }
+  }
+
+  Future<void> goToHallSVG() async { // May need return type of Future<void>, not sure yet
+    CameraPosition _currentPos = CameraPosition(
+        bearing: 123.31752014160156,
+        target: LatLng(45.49726709926478, -73.57894677668811),
+        tilt: 0.0,
+        zoom: 19.03557586669922);
+
+    final GoogleMapController controller = await _completer.future;
+    controller.animateCamera(CameraUpdate.newCameraPosition(_currentPos));
   }
 }
